can now cancel bookings

+++ graphql/resolvers/booking.js
@@ -3,7 +3,8 @@ const {
     findBookingData, 
     deleteBooking,
     findAllBookings,
-    createNewBooking
+    createNewBooking,
+    isValidBookingUser
 } = require("../../utils/booking");
 const { findEventData } = require("../../utils/event");
 
@@ -20,14 +21,19 @@ module.exports = {
             throw err;
         };
     },
+    bookingForUser: async({ userId }, req) => {
+        try {
+            if (!req.isUserAuthorized) throw new Error("User is unauthenticated!");
+        } catch(err) {
+            console.log(err);
+            throw err;
+        };
+    },
     createBooking: async ({ eventId, userId }, req) => {
         try {
             if (!req.isUserAuthorized) throw new Error("User is unauthenticated!");
-
-            console.log("Booking with user:", userId);
+            
             const booking = await createNewBooking(userId, eventId);
-
-            console.log("booking created!")
 
             await booking.save();
 
@@ -38,11 +44,18 @@ module.exports = {
             throw err;
         };
     },
-    cancelBooking: async ({ bookingId }, req) => {
+    cancelBooking: async ({ bookingId, userId }, req) => {
         try {
             if (!req.isUserAuthorized) throw new Error("User is unauthenticated!");
             
             const booking = await findBookingData(bookingId);
+
+            // make sure booking really exists
+            if (!booking || booking && !booking.user || booking && booking.user && !booking.user._id) throw new Error("Booking with that id does not exist!");
+
+            // make sure person deleting booking is the same one who created it
+            if(!isValidBookingUser(userId, booking.user._id.toString())) throw new Error("You can only delete your own bookings.");
+
             const event = await findEventData(booking.event._id);
 
             await deleteBooking(bookingId);
